pyproject.toml
[tool.poetry]
name = "discordBot"
version = "0"
description = "discordBot"
authors = ["discordBot"]

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^1.7.3"
Flask = "^2.1.2"
youtube_dl = "^2021.12.17"
PyNaCl = "^1.5.0"
replit = "^3.2.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
